[BPF] Allow to use no_caller_saved_registers for BPF
This commit allows to use no_caller_saved_registers to declare
BPF functions that do not clobber R0-R5 (caller saved registers).
For such functions do register allocation as-if caller saved registers
are not clobbered, but later wrap the calls with spill and fill
patterns that are simple to recognize in kernel.
Such scheme allows to generate BPF that:
- is compatible with old kernels that don't support functions with
  no_caller_saved_registers;
- allows new kernels that do support no_caller_saved_registers to
  remove unnecessary spill/fill instructions with little processing.

E.g. for the following C code:

   #define __no_csr __attribute__((no_caller_saved_registers))

   void bar(void) __no_csr;
   void buz(long i, long j, long k);

   void foo(long i, long j, long k) {
     bar();
     buz(i, j, k);
   }

First allocate registers as if:

   foo:
     call bar    # note: no spills for i,j,k (r1,r2,r3)
     call buz
     exit

And later insert spills fills on the peephole phase:

   foo:
     *(u64 *)(r10 - 8) = r1;  # Such call pattern is
     *(u64 *)(r10 - 16) = r2; # correct when used with
     *(u64 *)(r10 - 24) = r3; # old kernels.
     call bar
     r3 = *(u64 *)(r10 - 24); # But also allows new
     r2 = *(u64 *)(r10 - 16); # kernels to recognize the
     r1 = *(u64 *)(r10 - 8);  # pattern and remove spills/fills.
     call buz
     exit

The offsets for generated spills/fills are picked as minimal stack
offsets for the function. Allocated stack slots are not used for any
other purposes, in order to simplify in-kernel analysis.